+++ src/modules/esm.mjs
@@ -1,0 +1,43 @@
+import path, { dirname } from "node:path";
+import { release, version } from "os";
+import { createServer as createServerHttp } from "http";
+import "./files/c.js";
+import * as fs from "node:fs/promises";
+import { fileURLToPath } from "node:url";
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const random = Math.random();
+
+export let unknownObject;
+
+if (random > 0.5) {
+  await fs.readFile(path.join(__dirname, "files/a.json")).then((data) => {
+    unknownObject = JSON.parse(data);
+  });
+} else {
+  await fs.readFile(path.join(__dirname, "files/b.json")).then((data) => {
+    unknownObject = JSON.parse(data);
+  });
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+export const myServer = createServerHttp((_, res) => {
+  res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log("To terminate it, use Ctrl+C combination");
+});
